backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.19.0",
    "azure-monitor-opentelemetry>=1.6.4",
    "azure-storage-blob>=12.24.0",
    "fastapi>=0.115.6",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
]
